add homework, pyright config

+++ homework/celsius_to_fahrenheit.py
@@ -1,6 +1,6 @@
 from utils.input_utils import expect_for_valid_float_input
 
-celsius:int = expect_for_valid_float_input(-60)
+celsius:float = expect_for_valid_float_input(-60)
 
 fahrenheit:float = (float(celsius) * 9/5) + 32
